refactor(serve): move GPU memory config and queries into their own module
`engine/mod.rs` carried the `MemoryConfig` parser and the VRAM usage
queries inline, which is orthogonal to scheduling and kept growing as
backends were added. Lift both into `engine::memory` and re-export
`MemoryConfig` so callers are unaffected.

The limit parser now also accepts `GB`/`GiB`/`MiB` suffixes (all binary
units) rather than only the bare `G`/`M` forms, and the usage query has a
ROCm path alongside the CUDA one.
